qml: allow the settings component to load a dynamic action item.
This extends the usability of the settings component so that it is able 
to load any action item. This means that this component can be used for
all settings regardless of what action is meant to be on the right side.
This prevents the need for endless mutations on the settings component
to accomodate for different use-cases; the information component being
one such mutation.

Rebased-From: 40d48e7345b6af8ab6efe9f1869f25b3056290cc
Github-Pull: bitcoin-core/gui-qml#143