qml: remove unused inSubPage bool
This is not used at all. This is a remnant from a time where the idea
was to have the navbar encapsulated within the Wizard control, and this
bool would dictate what navbutton to display. We have since moved to
each page provides its navbar and its navbuttons.

Rebased-From: 5120735a5057a6d09853d726182e84cc280b4cc4
Github-Pull: bitcoin-core/gui-qml#198