qml: entangle the About and Developer options pages
Currently, the parent SwipeView of the AboutOptions page needs to have
an id of introductions. We abuse the knowledge that in our codebase we
will have that available, but this limits when and where the
AboutOptions can be used. This entangles the About and Developer options
so that we can avoid this, and have these pages truly be reusable.

Rebased-From: 6ac40e8482748225f6e913be50a742ba420a74b2
Github-Pull: bitcoin-core/gui-qml#203