qml: remember user's peer sort preference
This sets the sort preference for the peers table in qsettings so that
it can be remembered. Also addresses an issue where the `id`
toggle button is always checked when you go back to the table even
though the table is not sorted by `id`.

Rebased-From: 9cd558e17a9a21c979dae9fa7274bf066d15effe
Github-Pull: bitcoin-core/gui-qml#371