qml: Introduce DesktopWallets page
This is a main page when the AppMode is Desktop with walletEnabled.
The page contains a tab based layout with the sub pages being
selectable by clicking on a NavigationTab in the NavigationBar
at the top of the page.

Rebased-From: 4b79ac2c6e3c517268474444baf4f0b9519ecda1
Github-Pull: bitcoin-core/gui-qml#389