qml: Introduce the WalletBadge component
The WalletBadge is an Item that shows a wallet's name, type and
balance. The WalletBadge balance shows all of the decimal places
for a whole bitcoin and highlights the satoshis available. The
component can be configured to not show the Icon or the Balance.

Rebased-From: 5906e69bc13b0e7607a7fae7fa6e97ff624dc887
Github-Pull: bitcoin-core/gui-qml#401