qml: Introduce LabeledTextInput control
This control is primarily used in wallet forms where the user
is expected to input values for sending or receiving transactions.
It contains a Label describing the input as well as a TextInput
where the actual string is entered. A Icon is also available on
the right side of the component for extra actions.

Rebased-From: a3958386187e63dc6049451f76ed248012ff50ee
Github-Pull: bitcoin-core/gui-qml#435