send: access declared property color
This is setting a property in an Icon which is defined by us in
Icon.qml as a Button with required property color and
    icon.color: root.color.
In Qt6, only declared public properties are accessible from outside
the component. Qt5 was more lenient about this.

Rebased-From: 7fe66360d1bdb17a3a597c3cfefccdeb77f019cc
Github-Pull: bitcoin-core/gui-qml#475